third_party/blink/web_tests/fast/dom/Document/createElement-valid-names-expected.txt

Test various valid name for elements according to the definition of DOM Level 2.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement('ascii') is non-null.
PASS document.createElement('Ascii') is non-null.
PASS document.createElement('àscii') is non-null.
PASS document.createElement('Àscii') is non-null.
PASS document.createElement('āscii') is non-null.
PASS document.createElement('Ăscii') is non-null.
PASS document.createElement('ƻscii') is non-null.
PASS document.createElement('אscii') is non-null.
PASS document.createElement('āscii') is non-null.
PASS document.createElement('āscii') is non-null.
PASS document.createElement(':ascii') is non-null.
PASS document.createElement('_ascii') is non-null.
PASS document.createElement('aàscii') is non-null.
PASS document.createElement('aÀscii') is non-null.
PASS document.createElement('aāscii') is non-null.
PASS document.createElement('aĂscii') is non-null.
PASS document.createElement('aƻscii') is non-null.
PASS document.createElement('aאscii') is non-null.
PASS document.createElement('aāscii') is non-null.
PASS document.createElement('aāscii') is non-null.
PASS document.createElement('aa֑') is non-null.
PASS document.createElement('a⃣') is non-null.
PASS document.createElement('aि') is non-null.
PASS document.createElement('aʳ') is non-null.
PASS document.createElement('a0') is non-null.
PASS document.createElement('a---i') is non-null.
PASS document.createElement('a...i') is non-null.
PASS successfullyParsed is true

TEST COMPLETE
